+++ src/component/TodoInput.js
@@ -1,0 +1,20 @@
+export default function TodoInput({ addTodo }) {
+  const $input = document.querySelector(".new-todo");
+
+  const handleAddTodo = ({ target, key }) => {
+    if (key !== "Enter") {
+      return;
+    }
+
+    const contents = target.value.trim();
+    if (contents === "") {
+      return;
+    }
+
+    addTodo(contents);
+
+    target.value = "";
+  };
+
+  $input.addEventListener("keypress", handleAddTodo);
+}
